#4201 Move wallet connect and logout to Profile screen
