+++ packages/wallet/app/components/home/HomeScreen.tsx
@@ -1,35 +1,19 @@
-import { nonNullable } from "@neufund/shared-utils";
 import { useNavigation } from "@react-navigation/native";
 import React from "react";
-import { SafeAreaView, Text, View } from "react-native";
+import { View } from "react-native";
 
 import { appRoutes } from "../../appRoutes";
-import { authModuleAPI } from "../../modules/auth/module";
-import { walletConnectModuleApi } from "../../modules/wallet-connect/module";
 import { appConnect } from "../../store/utils";
-import { silverLighter2 } from "../../styles/colors";
 import { Button, EButtonLayout } from "../shared/buttons/Button";
+import { SafeAreaScreen } from "../shared/Screen";
 
-type TDispatchProps = {
-  logout: () => void;
-  walletConnectDisconnect: (peerId: string) => void;
-};
+type TStateProps = {};
 
-type TStateProps = {
-  user: NonNullable<ReturnType<typeof authModuleAPI.selectors.selectUser>>;
-  walletConnectPeer: ReturnType<typeof walletConnectModuleApi.selectors.selectWalletConnectPeer>;
-};
-
-const HomeLayout: React.FunctionComponent<TDispatchProps & TStateProps> = ({
-  walletConnectPeer,
-  walletConnectDisconnect,
-  user,
-  logout,
-}) => {
+const HomeLayout: React.FunctionComponent<TStateProps> = () => {
   const navigation = useNavigation();
 
   return (
-    <SafeAreaView style={{ backgroundColor: silverLighter2, flex: 1 }}>
+    <SafeAreaScreen>
       <View>
         <Button
           layout={EButtonLayout.PRIMARY}
@@ -39,41 +23,12 @@ const HomeLayout: React.FunctionComponent<TDispatchProps & TStateProps> = ({
           Scan QR code
         </Button>
       </View>
-
-      <Text>Address: {user.userId}</Text>
-
-      <Text>Wallet connect peer: </Text>
-      {walletConnectPeer ? (
-        <>
-          <Text>{JSON.stringify(walletConnectPeer, undefined, 2)}</Text>
-          <Button
-            layout={EButtonLayout.PRIMARY}
-            onPress={() => walletConnectDisconnect(walletConnectPeer.id)}
-          >
-            Disconnect
-          </Button>
-        </>
-      ) : (
-        <Text>not connected</Text>
-      )}
-
-      <Button layout={EButtonLayout.TEXT} testID="dashboard.logout" onPress={logout}>
-        Logout
-      </Button>
-    </SafeAreaView>
+    </SafeAreaScreen>
   );
 };
 
-const HomeScreen = appConnect<TStateProps, TDispatchProps>({
-  stateToProps: state => ({
-    walletConnectPeer: walletConnectModuleApi.selectors.selectWalletConnectPeer(state),
-    user: nonNullable(authModuleAPI.selectors.selectUser(state)),
-  }),
-  dispatchToProps: dispatch => ({
-    walletConnectDisconnect: (peerId: string) =>
-      dispatch(walletConnectModuleApi.actions.disconnectFromPeer(peerId)),
-    logout: () => dispatch(authModuleAPI.actions.logout()),
-  }),
+const HomeScreen = appConnect<TStateProps>({
+  stateToProps: () => ({}),
 })(HomeLayout);
 
 export { HomeScreen };
