+++ packages/wallet/app/components/profile/ProfileScreen.tsx
@@ -1,13 +1,15 @@
 import { nonNullable } from "@neufund/shared-utils";
 import { useNavigation } from "@react-navigation/native";
 import React from "react";
-import { StyleSheet, View } from "react-native";
+import { StyleSheet, Text, View } from "react-native";
 import Config from "react-native-config";
 
 import { appRoutes } from "../../appRoutes";
 import { authModuleAPI } from "../../modules/auth/module";
+import { walletConnectModuleApi } from "../../modules/wallet-connect/module";
 import { appConnect } from "../../store/utils";
 import { spacingStyles } from "../../styles/spacings";
+import { Button, EButtonLayout } from "../shared/buttons/Button";
 import { EIconType } from "../shared/Icon";
 import { SafeAreaScreen } from "../shared/Screen";
 import { AddressShare } from "./AddressShare";
@@ -16,11 +18,22 @@ import { Menu, EMenuItemType } from "../shared/menu/Menu";
 
 type TStateProps = {
   authWallet: NonNullable<ReturnType<typeof authModuleAPI.selectors.selectAuthWallet>>;
+  walletConnectPeer: ReturnType<typeof walletConnectModuleApi.selectors.selectWalletConnectPeer>;
+};
+
+type TDispatchProps = {
+  logout: () => void;
+  walletConnectDisconnect: (peerId: string) => void;
 };
 
 type TMenuProps = React.ComponentProps<typeof Menu>;
 
-const ProfileLayout: React.FunctionComponent<TStateProps> = ({ authWallet }) => {
+const ProfileLayout: React.FunctionComponent<TStateProps & TDispatchProps> = ({
+  authWallet,
+  walletConnectPeer,
+  walletConnectDisconnect,
+  logout,
+}) => {
   const navigation = useNavigation();
 
   const items = React.useMemo(() => {
@@ -51,6 +64,25 @@ const ProfileLayout: React.FunctionComponent<TStateProps> = ({ authWallet }) => 
       </View>
 
       <Menu items={items} />
+
+      <Text>Wallet connect peer: </Text>
+      {walletConnectPeer ? (
+        <>
+          <Text>{JSON.stringify(walletConnectPeer, undefined, 2)}</Text>
+          <Button
+            layout={EButtonLayout.PRIMARY}
+            onPress={() => walletConnectDisconnect(walletConnectPeer.id)}
+          >
+            Disconnect
+          </Button>
+        </>
+      ) : (
+        <Text>not connected</Text>
+      )}
+
+      <Button layout={EButtonLayout.TEXT} testID="dashboard.logout" onPress={logout}>
+        Logout
+      </Button>
     </SafeAreaScreen>
   );
 };
@@ -65,9 +97,15 @@ const styles = StyleSheet.create({
   },
 });
 
-const ProfileScreen = appConnect<TStateProps>({
+const ProfileScreen = appConnect<TStateProps, TDispatchProps>({
   stateToProps: state => ({
     authWallet: nonNullable(authModuleAPI.selectors.selectAuthWallet(state)),
+    walletConnectPeer: walletConnectModuleApi.selectors.selectWalletConnectPeer(state),
+  }),
+  dispatchToProps: dispatch => ({
+    walletConnectDisconnect: (peerId: string) =>
+      dispatch(walletConnectModuleApi.actions.disconnectFromPeer(peerId)),
+    logout: () => dispatch(authModuleAPI.actions.logout()),
   }),
 })(ProfileLayout);
